\name{mixreg-internal}
\alias{aux1}
\alias{aux2}
\alias{aux3}
\alias{boot.sam}
\alias{dir.sum}
\alias{gfun}
\alias{hatfun}
\alias{info1}
\alias{info2}
\alias{init.rand}
\alias{simmix}

\title{Internal mixreg functions}
\description{
  Internal mixreg functions.
}
\usage{
aux1(gma,k,l)
aux2(t11,t12,t13,t21,t22,t23,t31,t32,t33)
aux3(m,ind)
boot.sam(mu,resid,prob)
dir.sum(...)
gfun(x,y,theta)
hatfun(x,g)
info1(x,y,theta,gma)
info2(x,y,theta,gma)
init.rand(x,y,K,intercept)
simmix(theta,intercept,x) 
}

\details{
  These are usually not to be called by the user.
}
\keyword{internal}
