\name{aphids}
\alias{aphids}
\title{
Data on the rate of infection of tobacco plants by a virus
spread by aphids.
}
\description{
The \code{aphids} data frame has 51 rows and 2 columns.
These correspond to 51 independent experiments in which varying
numbers of aphids were released in a flight chamber containing
12 infected and 69 healthy tobacco plants.  The resulting number
of infected plants was recorded.
}
\usage{
data(aphids)
}
\format{
This data frame contains the following columns:
\describe{
\item{\code{n.aphids}}{
the number of aphids released in the flight chamber in
each instance
}
\item{\code{n.inf}}{
the resulting number (out of a possible 69) of infected plants
}}}
\source{
These data appear courtesy of Gilles Boiteau and George Tai of the
Potato Research Centre, Agriculture and Agri-Food Canada,
Fredericton, New Brunswick.  Any published work using these data
should cite the paper given in the \bold{References}.
}
\references{
Boiteau, G., M. Singh, R. P. Singh, G. C. C. Tai, and T. R. Turner
(1998).  Rate of spread of PVY-n by alate Myzus persicae (Sulzer)
from infected to healthy plants under laboratory conditions.  Potato
Research, vol. 41, pp. 335 -- 344.
}
\section{DETERMINATION OF INFECTION}{
After 24 hours, the flight chamber was fumigated to kill the aphids, and the
previously healthy plants were moved to a greenhouse and monitored to
detect symptoms of infection.  The number of plants displaying such
symptons was recorded.
}
\keyword{datasets}
% Converted by Sd2Rd version 1.21.
